+++ CHANGES.md
@@ -2,6 +2,23 @@
 
 all changes from the project are listed here. The changes are sorted by version
 
+## 3
+- new version for workfront update
+- include new styling
+- remove not used code
+- Remove old Options
+- include new icons
+- include check for timesheet
+- include all fine for your timesheet
+- include check for old sheets
+- include warning message as popup
+- include loading spinner
+- include mouse event listener for Button 3
+- work on design
+- include event listener for links on timesheets and timesheet
+- include fallbacks for loading file
+
+
 ## 2.2.1.1
 - HOTFIX for Loading Spinner
 - Change Colors from Loading Spinner